allow GetSettingsRequest to explicitly request relationships with unacknowledged users
Previously, if a user visited a prediction page whose creator trusted them, but who hadn't been marked trusted in return, the UI would claim that there was _no_ relationship, rather than an unreciprocated on. This fixes that, by forcing the relationship with the creator to be included in the returned settings.